Add dry_run variable

+++ Setup.ps1
@@ -1,12 +1,17 @@
+# This script serves:
+# - Download and unzip large contents.
+# - Download and build third-party libraries.
 
 #
 # Script arguments
 #
-
 Param (
 	[Switch]$skipdownload,
 	[Switch]$skipbuild
 )
+
+# For test
+$dry_run = $false
 
 $should_download = !($PSBoundParameters.ContainsKey('skipdownload'))
 $should_build = !($PSBoundParameters.ContainsKey('skipbuild'))
@@ -29,36 +34,49 @@ $contents_list  = @(
 #
 # Utilities
 #
+function Not-A-Drill {
+	return $dry_run -eq $false
+}
 function Download-URL {
 	Param ($webclient, $target_url, $target_path)
 	if (Test-Path $target_path -PathType Leaf) {
-		Write-Host $target_path, "already exists, will be skipped." -ForegroundColor Green
+		Write-Host "   ", $target_path, "already exists, will be skipped." -ForegroundColor Green
 	} else {
-		#Write-Host "url: ", $target_url -ForegroundColor Green
-		#Write-Host "target: ", $target_path -ForegroundColor Green
-		#Write-Host "Downloading..." -ForegroundColor Green
-		$webclient.DownloadFile($target_url, $target_path)
-		#Write-Host "Download finished."
+		if (Not-A-Drill) {
+			#Write-Host "url: ", $target_url -ForegroundColor Green
+			#Write-Host "target: ", $target_path -ForegroundColor Green
+			#Write-Host "Downloading..." -ForegroundColor Green
+			$webclient.DownloadFile($target_url, $target_path)
+			#Write-Host "Download finished."
+		}
 	}
 }
 function Clear-Directory {
 	Param ($dir)
 	Write-Host "Clear directory:", $dir -ForegroundColor Green
-	if (Test-Path $dir) {
-		Remove-Item -Recurse -Force $dir
+	if (Not-A-Drill) {
+		if (Test-Path $dir) {
+			Remove-Item -Recurse -Force $dir
+		}
+		New-Item -ItemType Directory -Force -Path $dir
 	}
-	New-Item -ItemType Directory -Force -Path $dir
 }
 function Ensure-Subdirectory {
 	Param ($dir)
-	New-Item -ItemType Directory -Path $dir -Force
+	Write-Host "Ensure directory:", $dir
+	if (Not-A-Drill) {
+		New-Item -ItemType Directory -Path $dir -Force | Out-Null
+	}
 }
 function Unzip {
 	Param ($zip_filepath, $unzip_dir)
-	if (Test-Path $unzip_dir) {
-		Remove-Item -Recurse -Force $unzip_dir
+	Write-Host "    Unzip to:", $unzip_dir
+	if (Not-A-Drill) {
+		if (Test-Path $unzip_dir) {
+			Remove-Item -Recurse -Force $unzip_dir
+		}
+		Expand-Archive -Path $zip_filepath -DestinationPath $unzip_dir
 	}
-	Expand-Archive -Path $zip_filepath -DestinationPath $unzip_dir
 }
 
 #
@@ -66,10 +84,10 @@ function Unzip {
 #
 $num_contents = $contents_list.length
 if ($should_download) {
+	Write-Host "Download contents... (count=$num_contents)" -ForegroundColor Green
+	
 	Ensure-Subdirectory "$content_dir"
 	$webclient = New-Object System.Net.WebClient
-	
-	Write-Host "Download contents... (count=$num_contents)" -ForegroundColor Green
 	
 	foreach ($desc in $contents_list) {
 		$content_url = $desc[0]
@@ -93,11 +111,6 @@ if ($should_download) {
 <#
 $freeimage_url      = "http://downloads.sourceforge.net/freeimage/FreeImage3180.zip"
 $freeimage_path     = "$external_dir/FreeImage.zip"
-$tinyobjloader_url  = "https://github.com/syoyo/tinyobjloader/archive/v2.0.0-rc1.zip"
-$tinyobjloader_path = "$external_dir/tinyobjloader-v2.0.0-rc1.zip"
-$content_url        = "https://casual-effects.com/g3d/data10/research/model/bedroom/bedroom.zip"
-$content_zip_path   = "$external_dir/content/bedroom.zip"
-$content_unzip_path = "$content_dir/bedroom/"
 
 #
 # Find MSBuild.exe and devenv.exe
@@ -114,16 +127,6 @@ Write-Host "devenv.exe:", $devenv_path -ForegroundColor Green
 # Build third party libraries
 #
 if ($should_build) {
-	### Content
-	if (Test-Path "$content_dir/bedroom") {
-		Write-Host "Content 'bedroom' already unzipped and will be skipped."
-	} else {
-		Ensure-Subdirectory "$content_dir/bedroom"
-		Write-Host "Unzip content 'bedroom'"
-		Expand-Archive -Path $content_zip_path -DestinationPath $content_unzip_path
-		Write-Host "Unzip done"
-	}
-	
 	### FreeImage
 	# 1. Unzip
 	if (Test-Path "$external_dir/FreeImage") {
@@ -148,22 +151,6 @@ if ($should_build) {
 	Copy-Item "./external/FreeImage/Dist/x64/FreeImage.h" -Destination $freeimage_source_dir
 	Copy-Item "./external/FreeImage/Dist/x64/FreeImage.dll" -Destination $freeimage_binary_dir
 	Copy-Item "./external/FreeImage/Dist/x64/FreeImage.lib" -Destination $freeimage_binary_dir
-	
-	### tinyobjloader
-	# 1. Unzip
-	if (Test-Path "$external_dir/tinyobjloader-2.0.0-rc1") {
-		Write-Host "tinyobjloader already unzipped. skip unzip."
-	} else {
-		Write-Host "Unzip tinyobjloader"
-		Expand-Archive -Path $tinyobjloader_path -DestinationPath $external_dir
-		Write-Host "Unzip done"
-	}
-	# 2. Copy
-	$tinyobjloader_source_dir = "./thirdparty/tinyobjloader/source"
-	Clear-Directory $tinyobjloader_source_dir
-	Write-Host "Copy source to $tinyobjloader_source_dir"
-	Copy-Item "./external/tinyobjloader-2.0.0-rc1/tiny_obj_loader.h" -Destination $tinyobjloader_source_dir
-	Copy-Item "./external/tinyobjloader-2.0.0-rc1/tiny_obj_loader.cc" -Destination $tinyobjloader_source_dir
 }
 
 Write-Host "> Done" -ForegroundColor Green
